Measure Gap plugin (#42)
* Measure Gap plugin

Allows a player to measure a gap between 2 points by using the menu

* Fixed to be compatible with old compiler

Now works with the Sourcemod 1.6 compiler, Beam color changed to cyan.

* v1.1, pointer beams.

Added pointer beams to show where the measure points are set to.  Sometimes the measure beam itself wont display, especially when snapping to grid and end points are inside geometry.  Pointer beams added for clarity of where measure points are set to.